Split out more types
